+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+## Unreleased
+
+### Enhancements
+
+* Allow a `Method` or any object responding to `#call` to be used as a callback or middleware
+  | [#662](https://github.com/bugsnag/bugsnag-ruby/pull/662)
+
 ## v6.20.0 (29 March 2021)
 
 ### Enhancements