+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+## TBD
+
+### Fixes
+
+* Deliver when an error is raised in the block argument to `notify`
+  | [#660](https://github.com/bugsnag/bugsnag-ruby/pull/660)
+  | [aki77](https://github.com/aki77)
+
 ## v6.20.0 (29 March 2021)
 
 ### Enhancements